test: share secure client defaults mixin
